+++ R/shaft-simple.R
@@ -13,13 +13,19 @@
 #'
 #' @param formatted An object coercible to [character].
 #' @param align Alignment of the column.
-#' @param na_indent Indention of `NA` values.
+#' @param na String to use as `NA` value, defaults to `"NA"` styled with
+#'   [style_na()] with fallback if color is not available.
+#' @param na_indent Indentation of `NA` values.
 #' @export
 #' @rdname new_pillar_shaft
 new_pillar_shaft_simple <- function(formatted, ..., width = NULL, align = "left",
-                                    min_width = NULL, na_indent = 0L) {
+                                    min_width = NULL, na = NULL, na_indent = 0L) {
   if (is.null(width)) {
     width <- get_max_extent(as.character(formatted))
+  }
+
+  if (is.null(na)) {
+    na <- pillar_na()
   }
 
   new_pillar_shaft(
@@ -28,6 +34,7 @@ new_pillar_shaft_simple <- function(formatted, ..., width = NULL, align = "left"
     width = width,
     min_width = min_width,
     align = align,
+    na = na,
     na_indent = na_indent,
     class = "pillar_shaft_simple"
   )
@@ -41,7 +48,11 @@ format.pillar_shaft_simple <- function(x, width, ...) {
   if (width < desired_width) {
     shaft <- str_trunc(shaft, width)
   }
-  shaft[is.na(shaft)] <- paste0(strrep(" ", attr(x, "na_indent")), pillar_na())
+  shaft[is.na(shaft)] <- paste0(
+    strrep(" ", attr(x, "na_indent")),
+    # Fallback for compatibility, remove pillar_na() call in pillar 1.5.0
+    attr(x, "na") %||% pillar_na()
+  )
 
   new_ornament(shaft, width = width, align = align)
 }