+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - `new_pillar_shaft_simple()` gains `na` argument to control appearance of `NA` values.
 
-- String columns where some but not all values need quotes are left-aligned visually. The values that don't require quotes gain an additional space for alignment (#171).
+- String columns are quoted if at least one value needs quotes (#171).
 
 - Apply subtle style to `list_of` columns (#172).